% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_get_spinup_op_sequence}
\alias{cbm_exn_get_spinup_op_sequence}
\title{get the default spinup op sequence names}
\usage{
cbm_exn_get_spinup_op_sequence()
}
\value{
list of strings - list of the names of the default spinup ops
}
\description{
get the default spinup op sequence names
}
