% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_get_step_disturbance_ops_sequence}
\alias{cbm_exn_get_step_disturbance_ops_sequence}
\title{get the default step op sequence names for disturbances}
\usage{
cbm_exn_get_step_disturbance_ops_sequence()
}
\value{
list of strings - list of the names of the default disturbance
step ops
}
\description{
get the default step op sequence names for disturbances
}
