% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_spinup}
\alias{cbm_exn_spinup}
\title{Run the cbm_exn spinup routine which initializes
C pools and state for simulations}
\usage{
cbm_exn_spinup(
  spinup_input,
  spinup_ops,
  spinup_op_list,
  parameters,
  spinup_debug_output_dir = NULL
)
}
\arguments{
\item{spinup_input}{a dictionary of spinup_parameters, spinup_increments}

\item{spinup_ops}{the formatted matrix operations to apply}

\item{spinup_op_list}{list of operation names to apply in spinup stepping
(references spinup ops)}

\item{parameters}{named list of default parameters for model initilization}

\item{spinup_debug_output_dir}{optional path which defaults to NULL.
If specified, the path is used to write debugging CSV files with full
model state and variable details about spinup timesteps.}
}
\value{
cbm_vars - a collection of simulation state and variables for
time-stepping with subsequent cbm processes
}
\description{
Run the cbm_exn spinup routine which initializes
C pools and state for simulations
}
