% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_spinup_ops}
\alias{cbm_exn_spinup_ops}
\title{Get the default spinup ops}
\usage{
cbm_exn_spinup_ops(spinup_input, parameters)
}
\arguments{
\item{spinup_input}{a dictionary of spinup_parameters, spinup_increments}

\item{parameters}{named list of default parameters for model initialization}
}
\description{
Get the default spinup ops
}
