% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_step}
\alias{cbm_exn_step}
\title{Run all C dynamics for one timestep}
\usage{
cbm_exn_step(
  cbm_vars,
  operations,
  disturbance_op_sequence,
  step_op_sequence,
  parameters
)
}
\arguments{
\item{cbm_vars}{named list of dataframes containing
the simulation state and variables for the timestep}

\item{operations}{list of formatted matrix operations
to apply to the pools.}

\item{disturbance_op_sequence}{list of named disturbance
operations to apply to the pools referencing the
values in the specified \code{operations}}

\item{step_op_sequence}{list of named annual process
operations to apply to the pools referencing the
values in the specified \code{operations}}

\item{parameters}{named list of default parameter for model
initialization}
}
\description{
Run all C dynamics for one timestep
}
