% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_step_ops}
\alias{cbm_exn_step_ops}
\title{Get the default matrix operations for cbm_exn stepping as a list of matrices}
\usage{
cbm_exn_step_ops(cbm_vars, parameters, cbm_exn_dir = NULL)
}
\arguments{
\item{cbm_vars}{named list of dataframes containing
the simulation state and variables for computing the
default operations}

\item{parameters}{named list of default parameters. For
default value see the \code{cbm_exn_get_default_parameters}
function.}

\item{cbm_exn_dir}{Path to a directory containing default parameters for \code{libcbm.cbm_exn.cbm_exn_model}}
}
\value{
list - list of structured matrix operations
}
\description{
Get the default matrix operations for cbm_exn stepping as a list of matrices
}
