% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_resources.R
\name{get_cbm_defaults_path}
\alias{get_cbm_defaults_path}
\title{get_cbm_defaults_path}
\usage{
get_cbm_defaults_path()
}
\description{
Get path of the CBM defaults SQLite database.
This will be the path to the packaged default database
or a path set with the option \code{libcbmr.cbm_defaults_path}.
}
