% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{get_cbm_exn_parameters_dir}
\alias{get_cbm_exn_parameters_dir}
\title{get_cbm_exn_parameters_dir}
\usage{
get_cbm_exn_parameters_dir()
}
\description{
Gets a path to a directory containing default parameters for \code{libcbm.cbm_exn.cbm_exn_model}
}
