% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{py_use_env}
\alias{py_use_env}
\title{Use conda or virtual environment}
\usage{
py_use_env(envname = NULL)
}
\arguments{
\item{envname}{Either the name of, or the path to, a Python virtual or conda environment.}
}
\description{
Will try to guess whether to use \code{use_condaenv()} or \code{use_virtualenv}.
}
