% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_get_default_parameters}
\alias{cbm_exn_get_default_parameters}
\title{Get the default parameters bundled with libcbm for running the cbm_exn
package}
\usage{
cbm_exn_get_default_parameters()
}
\description{
Get the default parameters bundled with libcbm for running the cbm_exn
package
}
