% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_exn.r
\name{cbm_exn_spinup_ops}
\alias{cbm_exn_spinup_ops}
\title{Get the default spinup ops}
\usage{
cbm_exn_spinup_ops(spinup_input, parameters, cbm_exn_dir = NULL)
}
\arguments{
\item{spinup_input}{a dictionary of spinup_parameters, spinup_increments}

\item{parameters}{named list of default parameters for model initialization}

\item{cbm_exn_dir}{Path to a directory containing default parameters for \code{libcbm.cbm_exn.cbm_exn_model}}
}
\description{
Get the default spinup ops
}
