% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_resources.R
\name{get_cbm_exn_parameters_dir}
\alias{get_cbm_exn_parameters_dir}
\title{get_cbm_exn_parameters_dir}
\usage{
get_cbm_exn_parameters_dir()
}
\description{
Get path of the parameters directory for \code{libcbm.cbm_exn.cbm_exn_model}.
This will be the path to the packaged default parameters
or a path set with the option \code{libcbmr.cbm_exn_parameters_dir}.
}
